Move some fixtures to functional/conftest.py
